requirements.txt
Django~=5.1.1
django-filter==22.1
django_debug_toolbar==3.8.1
djangorestframework~=3.15.2
ipython==8.8.0
psycopg2-binary==2.9.5
requests~=2.32.3
beautifulsoup4==4.12.3
python-dotenv==1.0.1